requirements.txt
Flask
requests
python-dotenv
flask-SQLAlchemy
flask_migrate
flask-APScheduler
python-dotenv